Factor out render_tool_catalog and tool_schemas_to_openai from bfcl.py + toolathlon.py into shared base.py environments module.
