+++ modal_training_gym/common/environments/toolathlon.py
@@ -41,6 +41,8 @@ from modal_training_gym.common.environments.base import (
     SandboxEnvironmentPool,
     StepResult,
     ToolCall,
+    render_tool_catalog as render_tool_catalog,
+    tool_schemas_to_openai as tool_schemas_to_openai,
 )
 
 # MCP servers whose state is confined to the agent workspace (snapshot-safe "Tier A").
@@ -860,67 +862,10 @@ train-ticket-plan.json
 Use the actual tool name, parameters, and full (untruncated) values required by your task; the call above is only an illustration of the wire format."""
 
 
-def render_tool_catalog(tool_schemas: dict, desc_chars: int = 160) -> str:
-    """Render the available-tools catalog the agent must choose from.
-
-    The agent can't query the MCP servers itself (that's the host loop's job via tools/list + the
-    OpenAI tools= param), so we surface the exact catalog the expert saw. Compact: one line per tool —
-    ``name(arg, required*)`` + a short description.
-    """
-    lines = []
-    for name in sorted(tool_schemas or {}):
-        spec = tool_schemas[name]
-        # tolerate both {description, parameters} and bare parameters (older cached rows)
-        if isinstance(spec, dict) and ("parameters" in spec or "description" in spec):
-            desc, params = spec.get("description", ""), spec.get("parameters", {})
-        else:
-            desc, params = "", spec
-        props = (params or {}).get("properties", {}) if isinstance(params, dict) else {}
-        required = (
-            set((params or {}).get("required", []) or [])
-            if isinstance(params, dict)
-            else set()
-        )
-        sig = ", ".join(f"{k}*" if k in required else k for k in props)
-        desc = " ".join(str(desc).split())[:desc_chars]
-        lines.append(f"- {name}({sig})" + (f": {desc}" if desc else ""))
-    return "\n".join(lines)
-
-
 def default_system_prompt(tool_schemas: dict) -> str:
     """Behavioral rules only — the tool catalog travels via the chat template's
     ``tools=`` parameter (see :func:`tool_schemas_to_openai`), not prompt text."""
     return DEFAULT_SYSTEM_PROMPT
-
-
-def tool_schemas_to_openai(tool_schemas: dict) -> list[dict]:
-    """Convert the dataset's ``{name: {description, parameters}}`` map into the
-    OpenAI/HF ``tools=`` list.
-
-    Pass the result to ``tokenizer.apply_chat_template(..., tools=...)`` (training
-    rollouts) or as the ``tools`` field of a chat-completions request (eval), so the
-    model sees the catalog in its NATIVE tool-calling format instead of a hand-rendered
-    text catalog, and emits calls in the wire format its template was trained on.
-    """
-    tools = []
-    for name in sorted(tool_schemas or {}):
-        spec = tool_schemas[name]
-        # tolerate both {description, parameters} and bare parameters (older cached rows)
-        if isinstance(spec, dict) and ("parameters" in spec or "description" in spec):
-            desc, params = spec.get("description", ""), spec.get("parameters", {})
-        else:
-            desc, params = "", spec
-        tools.append(
-            {
-                "type": "function",
-                "function": {
-                    "name": name,
-                    "description": desc,
-                    "parameters": params or {"type": "object", "properties": {}},
-                },
-            }
-        )
-    return tools
 
 
 def build_prefix_messages(
